curl_setup.h: merge back single-use internal header `curl_setup_once.h`
Originally split in 2006, but the issues cited are no longer present in
current code. As of now both `curl_setup.h` and `curl_setup_once.h` are
included once per compiler invocation, without recursion. The latter is
just a sub-header of the former with no clear distinction in their
contents.

Merge them to avoid having to decide where to put new global PP logic.
Also to make it easier to overview what gets defined globally and in
what order. (Perhaps even giving way of some tidying up here.)

Follow-up to 77b3bc239daf75d9fb7702ee34c8e5871c47d387
