+++ lib/Makefile.inc
@@ -306,7 +306,6 @@ LIB_HFILES =         \
   curl_rtmp.h        \
   curl_sasl.h        \
   curl_setup.h       \
-  curl_setup_once.h  \
   curl_sha256.h      \
   curl_sha512_256.h  \
   curl_share.h       \
